# api/openapi.yaml
openapi: 3.1.0
info:
  title: DokoSoko Control Plane API
  version: 4.0.0
  description: >-
    Control-plane contract for DokoSoko's core documentation, SDK references, credentials, recipes, reviewed tools, native tools, and MCP connector. Cookie-authenticated mutations
    also require CSRF validation. Public MCP is optional and read-only.
servers:
  - url: https://dokosoko.example.com
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/organisations:
    get:
      operationId: listOrganisations
      responses:
        '200':
          description: Organisation list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationList'
    post:
      operationId: createOrganisation
      requestBody:
        $ref: '#/components/requestBodies/NamedResource'
      responses:
        '201':
          description: Organisation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organisation'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/organisations/{organisation_id}/products:
    parameters:
      - $ref: '#/components/parameters/OrganisationID'
    get:
      operationId: listProducts
      responses:
        '200':
          description: Product list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
    post:
      operationId: createProduct
      requestBody:
        $ref: '#/components/requestBodies/NamedResource'
      responses:
        '201':
          description: Product created private-first with Public MCP disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
  /api/v1/products/{product_id}/description/rewrite:
    parameters:
      - $ref: '#/components/parameters/ProductID'
    post:
      operationId: rewriteProductDescription
      description: Uses the enabled assistant LLM profile to return an editable MCP-facing draft. The draft is never saved automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                draft:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required:
                - draft
      responses:
        '200':
          description: Unsaved rewritten description draft.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  description:
                    type: string
                    minLength: 1
                    maxLength: 1000
                required:
                  - description
        '422':
          $ref: '#/components/responses/Error'
...
  /api/v1/products/{product_id}/environments:
    parameters:
      - $ref: '#/components/parameters/ProductID'
    get:
      operationId: listEnvironments
      responses:
        '200':
          description: Environment list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentList'
    post:
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '201':
          description: Environment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dokosoko_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator automation token in development only; production uses the root session cookie.
    setupToken:
      type: http
      scheme: bearer
      description: One-time deployment setup token.
    dokoBearer:
      type: http
      scheme: bearer
      description: Opaque, short-lived DokoSoko OAuth access token bound to one exact Private MCP resource.
  parameters:
    ProductID:
      name: product_id
      in: path
      required: true
      schema:
        type: string
    OrganisationID:
      name: organisation_id
      in: path
      required: true
      schema:
        type: string
...
  requestBodies:
    NamedResource:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 120
              slug:
                type: string
                pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
            required:
              - name
              - slug
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    Organisation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        revision:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - revision
        - created_at
        - updated_at
    OrganisationList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organisation'
      required:
        - items
    ProductList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
      required:
        - items
    Environment:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        organisation_id:
          type: string
        product_id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        is_production:
          type: boolean
        revision:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - organisation_id
        - product_id
        - name
        - slug
        - is_production
        - revision
        - created_at
        - updated_at
    EnvironmentList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
      required:
        - items
    CreateEnvironmentRequest:
      type: object
      additionalProperties: false
      properties:
        organisation_id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        is_production:
          type: boolean
          default: false
      required:
        - organisation_id
        - name
        - slug
...
    Product:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        organisation_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          maxLength: 1000
        catalog_revision:
          type: integer
          format: int64
          minimum: 1
        public_mcp_enabled:
          type: boolean
          default: false
        revision:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - organisation_id
        - name
        - slug
        - description
        - catalog_revision
        - public_mcp_enabled
        - revision
        - created_at
        - updated_at
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - 'null'
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error
...